"pnpm-workspace.yaml // permite a instalação da biblioteca sharp para tratamento de image com nextjs, não vou mexer agora, mas vai ser útil futuramente
COPY pnpm-workspace.yaml ./ // adicionei o comando no dockerfile

alterei o globals.css // defini o estilo padrão para as telas, assim só vou usar ele como .css e os outros elementos serão definidos em seus módulos

criei a rota do login:
    - page.tsx recebe o modulo login-form;
    - layout.tsx editado para o estilo do login e tornei responsivo para maioria das telas mobiles obs: talvez precise de testes
criei a pasta components: // onde vai ficar os módulos criados
    - criei o modulo login-form // um formulário de login com nome e senha e uma função ainda não definida

deletei o frontend antigo // não vai mais ser necessário

OBS: definir a função para puxar a senha e o nome dos usuários da api
"
