--- app/frontend/old_react_frontend/src/app/assets/button.module.css
@@ -1,21 +1,0 @@
-.login {
-    border-style: none;
-    align-self: normal;
-    border-radius: 100px;
-    background-color: #A9A9A9;
-    font-size: x-large;
-    margin-top: 10%;
-    padding: 10px;
-    transition-duration: 0.5s;
-}
-.login:hover {
-    background-color: #bdbdbd;
-    box-shadow: 0px 0px 5px 2px black;
-    transition-duration: 0.5s;
-}
-.login:active {
-    background-color: #3a3a3a;
-    color: white;
-    box-shadow: 0px 2px 1px 3px black inset;
-    transition-duration: 0.5s;
-}
