--- app/frontend/old_react_frontend/src/app/globals.css
@@ -1,13 +1,0 @@
-@import "tailwindcss";
-
-html, body {
-  background-color: #A9A9A9;
-  font-family: "Inter", sans-serif;
-  font-size: x-large;
-  margin: 0;
-  padding: 0;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
