+++ app/frontend/src/app/login/layout.tsx
@@ -1,0 +1,24 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import "@/app/globals.css";
+
+const inter = Inter({subsets: ["latin"]});
+
+export const metadata: Metadata = {
+  title: "Login Page",
+  description: "Aba de Login do Atina",
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="pt-br">
+      <body className={`${inter.className} antialiased text-xl flex justify-center items-center`}>
+        {children}
+      </body>
+    </html>
+  );
+}
