+++ app/frontend/src/app/login/page.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+import Image from "next/image";
+import Login from "@/components/login-form";
+
+export default function Home() {
+  return (
+    <Login />    
+  );
+}
